Metric Schema and config name changes
In this commit, we are removing RemoteTarget and replacing with
RemoteResourceIdentifier and RemoteResourceType. Further, we are
formatting RemoteService, and the content of the RemoteResource
attributes such that they align with AWS Cloud Control resource names.

In addition to these changes, we are modifying the config names used for
enabling and configuring Application Signals to use the full feature name,
"application.signals".