Comment out E2E Operator Test (#940)
*Issue #, if available:*
E2E Operator test is currently failing due to a known issue. We do not
have an immediate plan to fix it so commenting it out for now.

Example failed run:
https://github.com/aws-observability/aws-otel-java-instrumentation/actions/runs/11525535146/job/32241628521

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.