Build layer during CI/CD workflows + some minor refactoring (#989)
*Description of changes:*
- Running the lambda layer build script during PR to ensure that layer
builds successfully.
- During the main build, ensure that we can build the lambda layer and
also upload it to the staging bucket for further E2E testing. The E2E
test will be added in the subsequent PR.

*Testing:*
Triggered the main_build workflow in my fork:
https://github.com/srprash/aws-otel-java-instrumentation/actions/runs/12679487319/job/35339366348
- Observed that the layer zip file was successfully uploaded to the run:
https://github.com/srprash/aws-otel-java-instrumentation/actions/runs/12679487319/job/35339366348#step:5:22
- Upload to S3 fails since I don't have the right credentials in my
fork.


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.