decouple apigw-lambda sample app from adot project build (#992)
The sample app doesn't need ADOT Javaagent instrumentation, rather it
depends on instrumentaiton via lambda layer. SO it makes sense to not
depend on adot javaagent build for building this sample app.


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.