fix packaging of the layer artifacts (#993)
The current logic is including the full path of the
`aws-opentelemetry-javaagent.jar` in the zip. But it should be at the
root along-with the `otel-instrument` file.

Tested locally and verified the directory structure is now correct.


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.