SigV4 Authentication support for http exporter (#1019)
Issue #, if available:
Adding SigV4 Authentication extension for Exporting traces to OTLP
CloudWatch endpoint without needing to explictily install the collector.

Description of changes:
Added a new class that extends upstream's OTLP http span exporter.
Overrides the export method so that if the endpoint is CW, we add an
extra step of injecting SigV4 authentication to the headers.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Co-authored-by: Mahad Janjua <majanjua@amazon.com>