feat: Support microservice span in Lambda Java environment.
When a microservice is hosted in a Lambda Java environment (e.g., a Spring Boot app), this PR ensures the following to provide a better observation experience:

The microservice will emit a server span that follows Lambda's server span.
Lambda's server span will be tagged with aws.trace.lambda.multiple-server=true.
If Lambda detects the HTTP call context, the operation attribute will be set as the HTTP method and path, instead of FunctionHandler.

Test:
All unit tests pass.
Built the Java layer and Spring Boot app on Lambda, and completed end-to-end testing.