Make the APIGW+Lambda sample app consistent with main branch (#1068)
The APIGW+Lambda Sample App in the `release/v1.33.x` branch has fallen
out-of-sync with the one in the `main` branch. This causes confusion
since we currently develop and release the Application Signals Lambda
Layer from the `release/v1.33.x` branch.


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.