Merge/v2.11.x (#1087)
*Issue #, if available:*

We have done the ADOT Java v2.11.0 and as part of the process, we need
to merge the changes we made for this release back to the main branch.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

*Description of changes:*
* Using GH UI created a new branch based on release/v.2.11.x
* Using GH UI created this PR to main
* Update otelSnapshotVersion to 2.12.0 in
`aws-otel-java-instrumentation.patch` and
`dependencyManagement/build.gradle.kts`
* Rebased from main

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Co-authored-by: Michael He <53622546+yiyuan-he@users.noreply.github.com>
Co-authored-by: ADOT Patch workflow <adot-patch-workflow@github.com>