Update Sonatype publishing URL to Central Portal (#1090)
It has been announced that the Sonatype OSSRH will reach EOL soon:
https://central.sonatype.org/news/20250326_ossrh_sunset/

We currently publish the ADOT Java and the UDP exporter to Sonatype and
the migration to Central Portal is recommended.

Sonatype publishing URL updated as per [this sonatype
doc](https://central.sonatype.org/publish/publish-portal-ossrh-staging-api/#configuration)
and the internal recommendation.


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Co-authored-by: Steve Liu <liustve@amazon.com>