update local operation of lambda span based on span attribute (#1106)
*Issue #, if available:*

*Description of changes:*

Currently the local operation of lambda span is hardcoded to
function_name//FunctionHandler. In some use cases, some server is
running inside lambda function and we should allow the setting of the
local operation dynamically at run time. For example, users can use span
api to set the attribute `aws.lambda.local.operation.override` and then
this value can be used for the local operation. We have an
Amazon-internal framework instrumentation that depends on this.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.