Base of AWS SDK v1.11 SPI Implementation (#1115)
This PR is similar to #1111, as it sets a base SPI implementation for
AWS SDK v1.11.

### Issue
This is the skeleton set up for the SPI, which aims to remove the
[patch](https://github.com/aws-observability/aws-otel-java-instrumentation/blob/main/.github/patches/opentelemetry-java-instrumentation.patch)
for aws-sdk v1.11 by using OTel's InstrumentationModule SPI extension.
This instrumentation is essentially complementing the upstream java
agent. It is completely separate from upstream and instruments after the
Otel agent.

### Description of Changes
This PR sets up the foundational structure for AWS SDK v1.11
Instrumentation, moving away from the current patching approach. It
doesn't modify current ADOT functionality or the upstream span. It just
registers the ADOT SPI implementation and sets up the interceptor hooks.

#### Core Components

1. **AdotAwsSdkInstrumentationModule**
   - Extends OpenTelemetry's `InstrumentationModule` SPI
- Registers custom handler through AdotAwsSdkClientInstrumentation class
in `typeInstrumentations` method

2. **AdotAwsSdkClientInstrumentation**
- AdotAwsSdkClientAdvice registers our handler only if the upstream
aws-sdk span is enabled (i.e. it checks if the upstream handler is
present when an AWS SDK client is initialized).

3. **AdotAwsSdkTracingRequestHandler**
   - Extends AWS SDK's `RequestHandler2`
   - Hooks into key SDK lifecycle points:
- `beforeRequest`: Captures final SDK request after upstream
modifications
- `afterAttempt`: Processes response before span closure in upstream
   - Will be used to enriches spans
   - Acts as central coordinator for all the awssdk_v1_11 components

4. **Resources Folder**
- Registers the v1.11 AdotAwsSdkInstrumentationModule into OTel's SPI
extension classpath in META-INF/services

### Key Design Decisions
1. **Instrumentation Ordering**
   - Deliberately structured to run after upstream OTel agent
   - Ensures all upstream modifications are captured
   - Maintains compatibility with existing instrumentation

2. **Lifecycle Hook Points**
   - `beforeRequest`: Last point to access modified request
   - `afterAttempt`: Final opportunity to enrich span before closure

### Testing
- Verified existing functionality remains unchanged and contract tests
pass (all contract tests pass after following the steps
[here](https://github.com/aws-observability/aws-otel-java-instrumentation/tree/main/appsignals-tests))
- Confirmed build success with new structure

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.