Send main build metrics (#1127)
*Issue #, if available:*

*Description of changes:*
Emit a failure metric if main build fails. Since this workflow is
triggered with pushes to main or a release branch, we want to be
notified if there is a failure with the build process or e2e tests.

Tested by temporarily adding an on: push: trigger to my own branch in
this repo and testing the updated workflow. Verified that failure metric
was successfully published to cloudwatch.

https://github.com/aws-observability/aws-otel-java-instrumentation/actions/runs/16783145872

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.