fix(deps): replace deprecated Gradle wrapper validation action (#1159)
The latest release of gradle/wrapper-validation-action is v3.5, but ADOT
Java is currently using v1.

This PR replaces the deprecated gradle/wrapper-validation-action with
the new gradle/actions/wrapper-validation action. v4.4.2 is an official
release. It's commit hash is #017a9ef.
https://github.com/gradle/actions/releases/tag/v4.4.2

When uses third-party action, we should always use commit hash instead
of release number, for security and integrity reason.
https://docs.github.com/en/actions/reference/security/secure-use#using-third-party-actions

References:
- https://github.com/gradle/wrapper-validation-action
- https://github.com/gradle/actions/releases/

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.