Add main build validation for release workflow (#1125)
*Issue #, if available:*

*Description of changes:*
This PR modifies the release build workflow to wait for the main build
workflow in the same branch to complete successfully. before proceeding
with the release. see [Python
PR](https://github.com/aws-observability/aws-otel-python-instrumentation/pull/443)
for more details and testing.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.