Update Kotlin version in UDP Exporter Sample Test App (#1184)
*Issue #, if available:*
- Fix failing workflow `udp-exporter-e2e-test`, where the sample app is
unable to be built
-
https://github.com/aws-observability/aws-otel-java-instrumentation/actions/runs/17587060722/job/49957841960

```
FAILURE: Build failed with an exception.

* What went wrong:
Could not determine the dependencies of task ':compileKotlin'.
> Unknown Kotlin JVM target: 21
```

*Description of changes:*
- Update Kotlin version in UDP Exporter Sample Test App. It is also now
aligned with the Kotlin version used in other `build.gradle.kts` files
in this repo.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.