Update Java lang version support from 22 to 23 (#1202)
We depend on [OTEL Java
2.11.0](https://github.com/aws-observability/aws-otel-java-instrumentation/blob/release/v2.11.x/dependencyManagement/build.gradle.kts#L30C20-L30C26),
which was released [Dec 23,
2024](https://github.com/open-telemetry/opentelemetry-java-instrumentation/releases/tag/v2.11.0),
at which Java 23 was supported (released [2024-09-17
](https://www.java.com/releases/)). We should have bumped this version
then, but we didn't have a good process in place at the time. Bump
version now.

Skipping changelog as we will support Java 24 soon.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.