feat: add self-validating workflow gate jobs (#1213)
Add gate jobs that fail if any workflow job fails OR if any job is
missing from the gate's needs array. Prevents both job failures and
configuration drift when adding new workflow jobs. Callout: I don't
think it's possible to have one gate for both workflows, but it should
not be the case that we add more over time.

### Testing:
See:
https://github.com/aws-observability/aws-otel-python-instrumentation/pull/477


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.