Update daily-scan.yml to run every 6 hours (#1240)
We still see periodic failures in this workflow and sometimes 2 failures
in a row is enough to cause an alarm (b/c of when the tests run and when
the alarm is evaluated). Increase run rate to every 6 hours to reduce
chance of 2 failures causing an alarm.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.