Release patch (#1255)
*Issue #, if available:*

*Description of changes:*
The 2.20.0 release build failed with at the step "Build and Publish
release with Gradle":

```
org.gradle.api.GradleException: version should not be set in build file when using nebula-release plugin. Instead use `-Prelease.version` parameter
```

The following PR fixes version.gradle.kts to not touch project.version
at all if the `-Prelease.version` flag is passed in to an environment.
It also creates a copy of the release build workflow that only re-runs
the failed Sonatype step in publish-sdk (did 2PR with @jj22ee to
confirm).

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.