Revert commented out release step (#1257)
*Issue #, if available:*

*Description of changes:*
In #1255, after making a fix to the version file, we re-ran the workflow
with the already-successful "Build and push image" step commented out.
This PR uncomments the step again for future patch releases on the
v2.20.x branch.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.